+++ src/effector/radarscan-effector.h
@@ -31,7 +31,7 @@ MGEFF_EFFECTOR effradarscaneffector_init(MGEFF_EFFECTOR _effector);
 void effradarscaneffector_finalize (MGEFF_EFFECTOR _effector);
 
 void effradarscaneffector_ondraw(MGEFF_ANIMATION animation,
-        MGEFF_EFFECTOR _effector, HDC sink_dc, int id, void* value);
+        MGEFF_EFFECTOR _effector, HDC sink_dc, intptr_t id, void* value);
 void effradarscaneffector_begindraw(MGEFF_ANIMATION animation,
         MGEFF_EFFECTOR _effector);
 void effradarscaneffector_enddraw(MGEFF_ANIMATION animation,
