+++ src/effector/star-effector.c
@@ -67,7 +67,7 @@ void effstareffector_finalize (MGEFF_EFFECTOR _effector)
 
 
 void effstareffector_ondraw (MGEFF_ANIMATION animation, MGEFF_EFFECTOR _effector,
-        HDC sink_dc, int id, void *value)
+        HDC sink_dc, intptr_t id, void *value)
 {
     EffEffector *effector = (EffEffector *)_effector;
     EffStarCtxt* star_context = (EffStarCtxt*)effector->context;
