+++ src/effector/star-effector.h
@@ -30,7 +30,7 @@ MGEFF_EFFECTOR effstareffector_init (MGEFF_EFFECTOR _effector);
 void effstareffector_finalize (MGEFF_EFFECTOR _effector);
 
 void effstareffector_ondraw (MGEFF_ANIMATION animation,
-        MGEFF_EFFECTOR _effector, HDC sink_dc, int id, void* value);
+        MGEFF_EFFECTOR _effector, HDC sink_dc, intptr_t id, void* value);
 
 void effstarscaneffector_begindraw (MGEFF_ANIMATION animation,
         MGEFF_EFFECTOR _effector);
