+++ src/effector/zip-effector.c
@@ -4,7 +4,7 @@
 
 #ifdef  _MGEFF_ZIPEFFECTOR
 void effzipeffector_ondraw (MGEFF_ANIMATION animation, MGEFF_EFFECTOR _effector,
-           HDC sink_dc, int id, void *value)
+           HDC sink_dc, intptr_t id, void *value)
 {
     EffEffector *effector = (EffEffector *)_effector;
     EffHDCSource *source1 = (EffHDCSource *) (effector->source_list.next);
